Added full compatibility with HF API

+++ bitmat/utils/rmsnorm.py
@@ -3,7 +3,7 @@ from ..triton_kernels.rmsnorm_kernel import fast_rms_layernorm
 
 
 class RMSLayerNorm(torch.nn.Module):
-    def __init__(self, shape, eps):
+    def __init__(self, shape, eps, dtype=torch.float16):
         super(RMSLayerNorm, self).__init__()
         self.eps = eps
         self.weight = torch.nn.Parameter(torch.randn(shape))
